requirements.txt
flask>=3.0
torch>=2.0
torchvision>=0.15
numpy>=1.24
Pillow>=10.0
imagehash>=4.3
scikit-learn>=1.3
tqdm>=4.65
ffmpeg-python>=0.2.0
